Refactored tests using helpers
