+++ test/helpers/empty-s3-dir.js
@@ -1,0 +1,19 @@
+module.exports = emptyS3Directory
+
+async function emptyS3Directory (s3, bucket, dir) {
+  const listParams = {
+    Bucket: bucket,
+    Prefix: dir
+  }
+  const listedObjects = await s3.listObjectsV2(listParams).promise()
+  if (listedObjects.Contents.length === 0) return
+  const deleteParams = {
+    Bucket: bucket,
+    Delete: { Objects: [] }
+  }
+  listedObjects.Contents.forEach(({ Key }) => {
+    deleteParams.Delete.Objects.push({ Key })
+  })
+  await s3.deleteObjects(deleteParams).promise()
+  if (listedObjects.IsTruncated) await emptyS3Directory(s3, bucket, dir)
+}
