+++ test/helpers/init.js
@@ -1,0 +1,7 @@
+const runAction = require('./run-action.js')
+const steps = ['pre', 'main']
+
+module.exports = () => {
+  runAction(steps)
+  return new Promise(resolve => setTimeout(resolve, 1500))
+}
